+++ docs/Whats-New.md
@@ -416,6 +416,7 @@ New:
 - [Event 606: AttachEffect is attaching to a Techno](AI-Scripting-and-Mapping.md#attacheffect-is-attaching-to-a-techno) (by FS-21)
 - [Linked superweapons](New-or-Enhanced-Logics.md#linked-superweapons) (by FS-21)
 - [Delayed fire weapons](New-or-Enhanced-Logics.md#delayed-firing) (by Starkku)
+- Burst without delay (by CrimRecya & TaranDahl)
 
 - New Missile trajectory (by CrimRecya)
 - New Engrave trajectory (by CrimRecya)